Rename module with test container
